# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cat Activities Monitor API",
    "description": "\n    A comprehensive FastAPI-based system for monitoring cats using YOLO object detection and activity analysis.    ",
    "license": {
      "name": "MIT License",
      "url": "https://github.com/biancarosa/cat-activities-monitor/blob/main/LICENSE"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/cats": {
      "get": {
        "tags": [
          "cats"
        ],
        "summary": "List Cat Profiles",
        "description": "List all cat profiles.",
        "operationId": "list_cat_profiles_cats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cats"
        ],
        "summary": "Create Cat Profile",
        "description": "Create a new cat profile.",
        "operationId": "create_cat_profile_cats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cats/{cat_name}": {
      "get": {
        "tags": [
          "cats"
        ],
        "summary": "Get Cat Profile",
        "description": "Get detailed information about a specific cat.",
        "operationId": "get_cat_profile_cats__cat_name__get",
        "parameters": [
          {
            "name": "cat_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cat Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cats"
        ],
        "summary": "Update Cat Profile",
        "description": "Update an existing cat profile.",
        "operationId": "update_cat_profile_cats__cat_name__put",
        "parameters": [
          {
            "name": "cat_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cat Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cats"
        ],
        "summary": "Delete Cat Profile",
        "description": "Delete a cat profile.",
        "operationId": "delete_cat_profile_cats__cat_name__delete",
        "parameters": [
          {
            "name": "cat_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cat Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cats/{cat_name}/activity-history": {
      "get": {
        "tags": [
          "cats"
        ],
        "summary": "Get Cat Activity History",
        "description": "Get activity history for a specific named cat.",
        "operationId": "get_cat_activity_history_cats__cat_name__activity_history_get",
        "parameters": [
          {
            "name": "cat_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cat Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CatActivity": {
        "type": "string",
        "enum": [
          "unknown",
          "sitting",
          "lying",
          "standing",
          "moving",
          "eating",
          "playing",
          "sleeping",
          "grooming"
        ],
        "title": "CatActivity",
        "description": "Enumeration of detectable cat activities.\n\nThese activities are recognized through pose analysis and behavioral patterns."
      },
      "CatProfile": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "Unique name identifier for the cat",
            "example": "Whiskers"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "General description or notes about the cat",
            "example": "A friendly orange tabby who loves to play with string toys"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Color",
            "description": "Primary color or color pattern of the cat",
            "example": "Orange tabby with white chest"
          },
          "breed": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Breed",
            "description": "Breed of the cat (if known)",
            "example": "Maine Coon"
          },
          "favorite_activities": {
            "items": {
              "$ref": "#/components/schemas/CatActivity"
            },
            "type": "array",
            "title": "Favorite Activities",
            "description": "List of activities this cat is commonly observed doing"
          },
          "created_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Created Timestamp",
            "description": "When this cat profile was first created",
            "example": "2023-12-01T14:30:22.123456"
          },
          "last_seen_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen Timestamp",
            "description": "When this cat was last detected in any camera",
            "example": "2023-12-01T16:45:30.789012"
          },
          "total_detections": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Detections",
            "description": "Total number of times this cat has been detected",
            "default": 0,
            "example": 127
          },
          "average_confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Average Confidence",
            "description": "Average detection confidence score for this cat",
            "default": 0.0,
            "example": 0.87
          },
          "preferred_locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Locations",
            "description": "Camera/image sources where this cat is commonly seen",
            "example": [
              "living_room_camera",
              "kitchen_camera"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "created_timestamp"
        ],
        "title": "CatProfile",
        "description": "Profile for a named cat with their characteristics.\n\nContains identification information, physical characteristics,\nbehavioral patterns, and detection statistics for individual cats."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}